requirements.txt
adjustText==1.0.4
albumentations==1.4.0
einops==0.7.0
hydra-core==1.3.2
hydra-colorlog==1.2.0
jinja2==3.1.3
matplotlib==3.8.3
numpy==1.26.4
opencv-python-headless==4.9.0.80
optuna==3.5.0
pandas==2.2.1
python-dotenv==1.0.1
pytorch_lightning==2.2.0.post0
pytest==8.0.2
PyYAML==6.0.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
SciencePlots==2.1.1
seaborn==0.13.2
tabulate==0.9.0
timm==0.9.16
torch>=2.2.1
torchmetrics==0.11.4
# torchvision==0.14.1
tqdm==4.66.2
wandb==0.16.3
